Tweak insert_tile_rand() and initial_board()
This is a preparation for the next patch: that patch will add an optimized way
of counting the number of empty tiles on the board. Though that new routine
doesn't work on a completely empty board. Luckily that situation only occurs
before the game has actually started in the initial_board() function and that
one is very easy to rewrite (don't use insert_tile_rand() for the very first
tile).

This patch also makes a small tweak in insert_tile_rand(). Unless there's a
bug in the game logic, it can never happen that insert_tile_rand() is called
on a board without at least one empty position. So we can remove the check for
that.