Changed find_best_move() -> do_best_move()
Before this patch the main loop looked (very roughly) like this:
  0) are there still legal moves?
  1) find best move
  1b) did step 1 really find a move?
  2) execute that move
  2b) was step 2 really a valid move?
But internally step 1 already executes that move, so the work in step 2 is
redundant. This patch changes step 1 to instead of the move (up, down, left,
right) return the resulting board of the move.

There were also a lot of redundant checks for valid moves. Step 0, 1b and 2b
are now collapsed into a single check.

This patch changes the interface of some of the exported functions. So it very
likely breaks the python bindings. I'm sure it's possible to update the python
stuff, but (for now) I didn't bother. Also the ask_for_move() function is
broken now (but can easily be fixed), but again I didn't do this.

@Robert Xiao:
The original game-loop was a more natural fit for interactivly playing the game
and likely also to automatically controlling a web-based game. This new code is
(only) better suited to quickly solve the game. So feel free to drop this patch
when integrating this branch into yours.