Cleanup init_xxx_table() routines
Both init_move_tables() and init_score_tables() fill in tables which have
0x10000 entries. Both routines contain some common operations. It's possible
to merge both into one. (Of course this won't be noticable in the total
execution speed).

This also changes the API of the .so file. Though IMHO it's better to only
expose one init() function to better hide the internal implementation details.